[UPD] Changed project's structure.
