+++ animators/cssProperties/options.ts
@@ -1,0 +1,16 @@
+import { BaseAnimatorOptions } from "../../animators";
+
+export default interface CssPropertyAnimatorOptions extends BaseAnimatorOptions
+{
+    name: string;
+    unit?: string;
+
+    startValue: number;
+    endValue?: number;
+
+    // direction?: string; -> normal / inverse
+    // speed?: number; -> 1x / 2x / 0.5x / ...
+    // timing?: string; -> linear / ease / cubic
+
+    computeValue?: (ratioValue: number) => number;
+}
