+++ animators/options.ts
@@ -1,0 +1,7 @@
+export default interface BaseAnimatorOptions
+{
+    target?: HTMLElement;
+
+    minWidth?: number;
+    maxWidth?: number;
+}
